pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chessgizmo"
version = "0.0.1"
description = "A library for chess data analysis, visualization and rating prediction."
authors = [
  { name = "tehnik-tehnolog", email = "mail@mail.mail" }
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "chess",
    "pandas",
    "stockfish",
    "berserk",
    "mureq",
    "sqlalchemy",
    "matplotlib",
    "numpy",
    "seaborn",
    "plotnine",
    "psycopg2-binary",
    "aioboto3",
    "python-dotenv",
    "scikit-learn==1.6.1",
    "cairosvg",
    "catboost"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["chessgizmo*"]

[tool.setuptools.package-data]
"chessgizmo" = ["data/**/*", "sql/**/*", "icons/**/*"]
#"chessgizmo" = ["**/*.pkl", "**/*.png", "**/*.sql", "**/*.json"]